+++ src/Exception/RouteAlreadyExistsException.php
@@ -4,12 +4,12 @@ declare(strict_types=1);
 
 namespace Fhooe\Router\Exception;
 
-use RuntimeException;
+use LogicException;
 
 /**
- * Exception thrown when attempting to add a route that already exists.
+ * Exception thrown when attempting to add a route that already exists (configuration error).
  * @package Fhooe\Router\Exception
  * @author Wolfgang Hochleitner <wolfgang.hochleitner@fh-hagenberg.at>
  * @since 2.0.0
  */
-class RouteAlreadyExistsException extends RuntimeException {}
+class RouteAlreadyExistsException extends LogicException {}